add policy() function to State

+++ contracts/PolicyVote.sol
@@ -4,14 +4,13 @@ import "openzeppelin-solidity/contracts/math/SafeMath.sol";
 
 contract PolicyVote {
 	using SafeMath for uint256;
-	mapping(address => uint) private voteNumber;
+	mapping(address => uint256) private voteNumber;
 
 	function vote(address policyAddress, uint256 vote) public {
 		voteNumber[policyAddress] = voteNumber[policyAddress] + vote;
 	}
 
 	function votingRsult() private returns (address) {
-
 		return address(0);
 	}
 }
@@ -24,6 +23,5 @@ contract PolicyVoteProvider {
 	function vote(address policyAddress, uint256 vote) public {
 		policyVote.vote(policyAddress, vote);
 	}
-
 
 }
